// typechain-types/src/DAOFactory/DaoFactory.ts
export declare namespace IDaoFactory {
  export type CreateParamsStruct = {
    daoName: PromiseOrValue<string>;
    owner: PromiseOrValue<string>;
    vote_maximumSupply: PromiseOrValue<BigNumberish>;
    vote_name: PromiseOrValue<string>;
    vote_symbol: PromiseOrValue<string>;
    vote_URI: PromiseOrValue<string>;
    timelock_minDelay: PromiseOrValue<BigNumberish>;
    governance_votingDelay: PromiseOrValue<BigNumberish>;
    governance_votingPeriod: PromiseOrValue<BigNumberish>;
    governance_quorumPercentage: PromiseOrValue<BigNumberish>;
  };

  export type CreateParamsStructOutput = [
    string,
    string,
    BigNumber,
    string,
    string,
    string,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    daoName: string;
    owner: string;
    vote_maximumSupply: BigNumber;
    vote_name: string;
    vote_symbol: string;
    vote_URI: string;
    timelock_minDelay: BigNumber;
    governance_votingDelay: BigNumber;
    governance_votingPeriod: BigNumber;
    governance_quorumPercentage: BigNumber;
  };
}